change index and  break up server (api rest) and socket
